chore: split ThanosStack into multiple modules
