+++ commands/update_network.go
@@ -3,8 +3,10 @@ package commands
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/tokamak-network/trh-sdk/pkg/constants"
+	"github.com/tokamak-network/trh-sdk/pkg/logging"
 	"github.com/tokamak-network/trh-sdk/pkg/stacks/thanos"
 	"github.com/tokamak-network/trh-sdk/pkg/types"
 	"github.com/tokamak-network/trh-sdk/pkg/utils"
@@ -37,6 +39,10 @@ func UpdateNetwork(ctx context.Context, network, stack string, config *types.Con
 		fmt.Println("You are using the local devnet. No need to update the network.")
 		return nil
 	}
+
+	// Initialize the logger
+	fileName := fmt.Sprintf("logs/update_network_%s_%s_%d.log", stack, network, time.Now().Unix())
+	logging.InitLogger(fileName)
 
 	switch stack {
 	case constants.ThanosStack:
